requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cloudpickle==1.2.1
cryptography==2.1.4
cycler==0.10.0
dask==2.1.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.6
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
idna==2.6
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-http-over-ws==0.0.6
jupyterlab==1.0.3
jupyterlab-server==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
locket==0.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
pandas==0.25.0
pandocfilters==1.4.2
parso==0.4.0
partd==1.0.0
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pyarrow==0.14.1
pycrypto==2.6.1
Pygments==2.4.2
pygobject==3.26.1
pyparsing==2.4.0
pyrsistent==0.15.2
python-apt==1.6.4
python-dateutil==2.8.0
pytz==2019.1
pyxdg==0.25
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.1
scikit-learn==0.21.2
scipy==1.3.0
SecretStorage==2.3.1
Send2Trash==1.5.0
six==1.11.0
tb-nightly==1.14.0a20190603
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019060501
toolz==0.10.0
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.1
